requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-generativeai==0.3.2
praw==7.7.1
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0

